+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `force` flag for `git push` command
 
+### Changed
+
+- `repo` to `name` flag name to mention repository names in `groups create`, `groups append`, `groups exclude` commands
+
 ## [0.9.0] - 2022-11-04
 
 ### Added